+++ aider/gui.py
@@ -9,7 +9,7 @@ import streamlit as st
 from aider import urls
 from aider.coders import Coder
 from aider.dump import dump  # noqa: F401
-from aider.io import InputOutput
+from aider.input_output import InputOutput
 from aider.main import main as cli_main
 from aider.scrape import Scraper
