+++ tests/basic/test_commands.py
@@ -12,7 +12,7 @@ import git
 from aider.coders import Coder
 from aider.commands import Commands, SwitchCoder
 from aider.dump import dump  # noqa: F401
-from aider.io import InputOutput
+from aider.input_output import InputOutput
 from aider.models import Model
 from aider.repo import GitRepo
 from aider.utils import ChdirTemporaryDirectory, GitTemporaryDirectory, make_repo
