+++ tests/basic/test_editblock.py
@@ -8,7 +8,7 @@ from unittest.mock import MagicMock, patch
 from aider.coders import Coder
 from aider.coders import editblock_coder as eb
 from aider.dump import dump  # noqa: F401
-from aider.io import InputOutput
+from aider.input_output import InputOutput
 from aider.models import Model
