+++ tests/basic/test_repo.py
@@ -9,7 +9,7 @@ from unittest.mock import patch
 import git
 
 from aider.dump import dump  # noqa: F401
-from aider.io import InputOutput
+from aider.input_output import InputOutput
 from aider.models import Model
 from aider.repo import GitRepo
 from aider.utils import GitTemporaryDirectory
