+++ tests/basic/test_repomap.py
@@ -2,7 +2,7 @@ import os
 import unittest
 
 from aider.dump import dump  # noqa: F401
-from aider.io import InputOutput
+from aider.input_output import InputOutput
 from aider.models import Model
 from aider.repomap import RepoMap
 from aider.utils import IgnorantTemporaryDirectory
