+++ tests/basic/test_wholefile.py
@@ -8,7 +8,7 @@ from unittest.mock import MagicMock
 from aider.coders import Coder
 from aider.coders.wholefile_coder import WholeFileCoder
 from aider.dump import dump  # noqa: F401
-from aider.io import InputOutput
+from aider.input_output import InputOutput
 from aider.models import Model
